JS: Functional Programming 2
We are talking about the default settings. As we remember, in JS, parameters that do not get an argument are initialized as "undefined". Sometimes we may not know if some values ​​will be passed, so we can set them as default. We can also pass methods or even a function call in default values. Arguments that do not have their own parameter are simply ignored.