JS: Functional Programming 3
Execution of functions. An expression is code that, when executed, returns a value. An instruction is a code representing a command. Expressions are calculated. Instructions are executed. Expressions include: function call, arithmetic and logical operations, ternary operator. Instructions include: for, while, break, return, if. Now we know that a function can be a parameter, since a function is an expression.