JS: Functional Programming 4
A few words about objects of the first class. Such an object can be any data that can be passed to a parameter, returned from a function, or assigned to a variable. The simplest are strings and numbers. A function is also a data type and we can assign it to a constant, as we often do. The main conclusion is that the definition of a function is an expression, which means it returns a value, namely a function. Since a function is an expression, then we can call it without a constant. In the task, 3 parameters are passed to us (the number of calls, the function used, the number is the argument). The terminal case occurs when the counter is 0, in which case the argument is returned to us. Next, we decrease the counter and run our function rucursively, passing the result of the second parameter (function) to the third parameter our argument.