JS: Functional Programming 5
Lexical environment lesson. Every time a function is called in the program, a special LexicalEnvironment dictionary (lexical environment) is created inside the interpreter associated with this call. All definitions of constants, variables, and other things inside the function are automatically written to the dictionary. The lexical environment is a store for in-memory data and a mechanism for retrieving that data on access. The dictionary is populated when the function is called, not when it is defined. During the execution of the function, the values ​​of variables can change, which is immediately reflected in the lexical environment. After the function is executed, its lexical environment is destroyed, and the memory occupied by it is freed. There is an exception to this behavior - the return of a function. An overlap is a situation where an identifier with the same name is created in the inner environment as in the outer one. And it does not matter what it is: a function argument, a constant or a variable.