JS: Functional Programming 6
Returning functions from functions. Since functions are the same data as strings and numbers, they can be passed to other functions as an argument. You can define a function inside another function and pass it outside. Anything to the right of => is considered the body of the function. The number of investments is not limited in any way.