Trial "Age Counter"
As an input, we get an array with users objects. You need to return an object with the properties of years of birth, and the value is the number of men born in this year. Only higher-order functions (filter, map, reduce) are used in the teacher's solution. I used only one and collected a new object through the loop.