JS: Composite Data 1
The next course, topic Composite data. Composite data is used in the following cases. When we write programs, we most often try to model quite complex processes and phenomena. When working with complex computational objects, thanks to composite data, as I said, we can raise the so-called conceptual level. Data abstraction is a method of separating the parts of a program that deal with the representation of data objects from the parts where those objects are used. With the help of several abstractions, in practice, we need to write a function for finding a quadrant, a function for mirroring points, and a function for finding the middle of a segment.