JS: Composite Data 2
Composite data, pairs. You can’t just take and change a couple. You can only create a new one based on the previous one. At first, this way of programming may seem unusual and difficult, as you need to restructure your thinking. A pair is only what is created using the cons constructor. If, for some reason, not a pair was passed to the selectors, then the result will be just such an error.