Trial " JSON stringify"
It is necessary to implement a function of the JSON stringify type, the difference is that the properties and values ​​returned in the string must be without quotes. Otherwise, everything is the same, so at the input we have a value, an indentation type and an indentation amount. We write another iterative function in which our value and depth, do a check for primitive data and null (since it is an object in js). In this case, we simply return our value. Now we have only objects left. We write three constants: indent size, current indent, bracket indent. Now we form arrays from our object through Object.entries and run the .map([key, val]) chain. Through interpolation we form a string: current indent, key, colon, iter function with val and depth +1. Now we return everything: the opening bracket, expand our values, indent the bracket, close the bracket, and use the join('\n') method. At the very end, we return our iterative function with the original value and depth one.