JS: Composite Data 4
Abstraction barriers. The concept is that, when working in one subject area on a certain slice, they operate with objects only of this slice, avoiding objects that do not belong to it. According to the assignment, we need to implement the function of creating a rectangle, consisting of the top left point, width and height, selectors, finding the area and perimeter, and a function that determines the location of the center of the axes in the rectangle.