JS: Introduction to OOP 7
This implementation uses a combination of function currying and object coercion to achieve the desired behavior. When the magic function is called with one or more arguments, it calculates the sum of those arguments and returns a new function f that takes another set of arguments. This process can be repeated as many times as necessary until the final call is made. To ensure that the final call returns the correct sum, we define a valueOf method on the function f that returns the sum calculated so far.

Packing and unpacking are two techniques used in JavaScript to convert between primitive values and objects.

Packing, also known as boxing, involves wrapping a primitive value in an object using a constructor function that corresponds to the primitive type. This can be useful when working with APIs that expect objects instead of primitives.

Unpacking, also known as unboxing, involves extracting the primitive value from an object. This is done using various methods and operators, such as valueOf, toString, and the unary plus operator.

While packing and unpacking can be useful in certain situations, they can also have performance implications due to the creation and destruction of objects. Therefore, it's important to use these techniques judiciously and only when necessary. The valueOf method is called automatically when a function is used in a context that expects a primitive value, such as when it is added to a number or passed as an argument to a function that expects a number. This allows us to use the magic function in arithmetic expressions and comparisons as shown in the tests.