JS: Introduction to OOP 8
The toString() method is a built-in prototype method in JavaScript that is used to convert an object to a string representation. It can be called on any object and returns a string that represents the object. By default, toString() returns a string representation of an object that includes the object's type and memory address. However, many built-in JavaScript objects and user-defined classes override the toString() method to provide a more meaningful string representation of the object. To override the toString() method for a user-defined class, the method must be defined within the class and return a string representation of the object in the desired format. When the toString() method is called on an instance of the class, the overridden method will be executed instead of the default method. The toString() method is commonly used to display object information to users, to serialize objects for storage or transmission, and to compare objects for equality.