JS: Introduction to OOP 9
The Cart class has four methods: The constructor creates an empty this.items array to hold the items. The addItem method adds items to the cart. If the product is already in the cart, then its quantity is increased by count, otherwise a new object is added to the items array. The default number is 1. The getItems method returns an array of objects with two properties: item is an item object and count is the number of items in the cart. The getCost method calculates the total cost of the cart by looping through the items array and summing the product of the price of the item and its quantity. The getCount method calculates the total number of items in the cart by iterating through the items array and summing the quantity of each item.
A class in JavaScript is a template or a blueprint for creating objects that share common properties and methods. Here are several theses describing classes in JavaScript:

A class is defined using the class keyword followed by the name of the class. A class can have a constructor method, which is called when an object is created from the class. The constructor is used to initialize the object's properties. Properties are defined using this keyword within the constructor or outside of it. Methods are defined within the class using function syntax. Methods can access and modify the properties of the class using the this keyword. Methods can be defined in the class declaration or added later using the prototype object. Objects can be created from a class using the new keyword followed by the name of the class and any ecessary arguments to the constructor method. Classes can be extended using the extends keyword to create a subclass that inherits properties and methods from the parent class. Subclasses can override methods or properties of the parent class using the same name to provide custom functionality. Classes can implement interfaces using the implements keyword, which requires the class to have the same methods as specified in the interface.