JS: Object Oriented Design 1
In object-oriented design in JavaScript, a configuration is a design pattern that is used to create objects with given parameters. Instead of passing many parameters to the object's constructor, we pass a configuration object with a set of properties that define the object's parameters. Configuration allows you to make objects more flexible and customizable, as well as reduce the number of arguments passed to the constructor. A configuration can be used to customize any object that can be customized, such as a user interface, application configuration, or class. Typically, a configuration is passed as an object with properties that can be changed by default. This object can then be passed to the object's constructor or method, which will use the configuration properties to configure the object. The configuration can contain any properties needed to customize the object. In an object's constructor, these properties can be stored as object properties, allowing them to be used later. If some configuration properties were not passed, the default values ​​are used.