JS: Object Oriented Design 2
A mutable configuration in JavaScript is an object that contains settings or options that can be changed at runtime. A configuration can contain various types of data such as strings, numbers, arrays, objects, and so on. According to the assignment, it is necessary to implement the constructor and the truncate() method in the Truncater class. The method accepts text and the following options:
separator - a character that replaces the trimmed part of the string
length - the maximum length of the source string. If the string is not longer than this option, then the original string is returned.