JS: Composite Data 6
Extreme Practice on Composite Data, tells how the concept of creating pairs fits well with the creation of rational numbers. According to the assignment, it was necessary to create such an abstraction (creation of a rational number constructor, selectors and actions that can be performed with fractions (addition, multiplication, division, subtraction).