Trial "Pairs on strings"
According to the assignment, you need to implement cons() - it takes a string that contains a certain character (it divides our string into two parts), car() - returns the first part of the string, cdr() - the second. The split method came up very well here, which split our string. But the teacher's decision is completely different, it shows the beauty of functional programming.