JS: Object Oriented Design 3
In JavaScript, objects can be used as entities (entity) and values ​​(value), as well as to implement embedded objects (embedded objects). Entity Objects are objects that represent a specific entity in the system. For example, a user object in an application. An entity object usually has properties that store information about the object, as well as methods that let you work with that information. Entity objects are usually created using constructors and are used to model the application's business logic. Value Objects are objects that represent a simple value, such as a date or money. The value object does not have its own identifier and is usually created each time it is needed. Value objects are usually created using factory methods or just functions that create a new value object each time they are called. Embedded Objects are objects that are part of other objects and are used to group properties and methods in one place. For example, a user object may contain nested objects such as an address object or a contact information object. Embedded objects are created inside other objects and can only be accessed from the context of that object.