JS: Object Oriented Design 4
Fluent Interface is a design pattern that allows you to write code that reads like natural language. It relies on methods that return objects, allowing them to be chained together, and on the use of object literals. The essence of the pattern is to create an object to which methods can be applied sequentially. As a result, each method modifies the object and returns it, which allows the next method to be applied to the already modified object.
According to the assignment, you need to implement the normalize () function, which takes a list of cities and countries as input, normalizes their names, sorts the cities and groups them by country.