JS: Object Oriented Design 5
Builders are design patterns that are used to create complex objects with parameters and settings. Instead of passing many arguments to an object's constructor, we use a collector to define the parameters one by one. In JavaScript, assemblers can be implemented as classes with method call chaining. Return methods of class instances, which allow chaining of method calls on a sequential parameter.