JS: Object Oriented Design 6
The JavaScript Proxy object is a generic wrapper for objects and provides the ability to intercept and change the behavior of the properties and methods of the object it refers to. Using the Proxy object, you can define traps for various operations, such as reading, writing and deleting object properties, calling its methods, etc. Proxy can also be used to create virtual objects that can access a remote object or cache data.