HTTP protocol 1
HTTP (HyperText Transfer Protocol) is a data transfer protocol that is used to transfer information between web servers and clients on the Internet. HTTP operates on a client-server model, where the client sends a request to the server, and the server responds to the request by sending back the requested information in the form of an HTML page, images, videos, and other resources. Every HTTP request has three main components: method, URI, and protocol version. The method specifies the request type, the URI specifies the requested resource, and the protocol version specifies the version of HTTP being used.