HTTP protocol 2
HTTP 1.1 is the latest version of the HTTP protocol, which has a number of improvements over the previous version (HTTP 1.0), such as: Persistent Connection: HTTP 1.1 supports a persistent connection between a client and a server, allowing the connection to be reused to send multiple requests and responses between the client and server.Caching: HTTP 1.1 includes a caching mechanism that allows clients to store local copies of resources such as pages and images on their computer to reduce download times and reduce server load. Additional request methods: HTTP 1.1 includes several additional request methods, such as PUT, DELETE, OPTIONS, TRACE, and CONNECT, which allow more flexible management of resources on the server. Extensibility: HTTP 1.1 was designed with the possibility of adding new extensions and features in the future, allowing the protocol to be extended. Parallel transmission: HTTP 1.1 allows multiple requests and responses to be transmitted in parallel between a client and a server within a single connection, which increases page loading speed. The HTTP 1.1 protocol uses a standard set of status codes such as 200 (OK), 404 (Not Found), and 500 (Internal Server Error) to inform the client of the status of a request and the result of the request being processed by the server. The client may also send request and response headers that contain additional information about the request and the result of its execution.In general, HTTP 1.1 is a widely used data transfer protocol on the Internet and provides high speed and reliability of data transfer between server and client.

Task: Using telnet make a request to hexlet.local (located on localhost) on port 8080. Request parameters: method GET, page /, protocol HTTP 1.1; Write your HTTP request to a solution file;