HTTP protocol 3
The HTTP request body is the data that is sent along with the request to the server. The request body can contain different data depending on the type of request and the purpose of the request. For example, for a POST request, the request body may contain data that needs to be added on the server, while for a GET request, there is usually no request body. The request body can be sent in various formats such as JSON, XML, HTML, text format, and others. In an HTTP request, the request body is specified after the request headers, separated by an empty string from the headers, and passed to the server.