HTTP protocol 4
Submitting forms in HTTP is done using the POST method. In this case, the form data is transmitted in the request body (request body), and not in the URL. In order to correctly submit the form, you need to set the correct headers. Each form field must have its own name (name) and value (value). When the form is submitted, the field values ​​are concatenated into a single string and passed in the request body in the application/x-www-form-urlencoded encoding. This replaces special characters (such as spaces and & or = signs) with the appropriate codes. You can use a regular HTML form element to submit a form in HTTP, or you can submit it with JavaScript using the XMLHttpRequest object or fetch.