HTTP protocol 5
Transfer-Encoding: chunked is a mechanism that allows you to transfer content as a sequence of several fragments (chunk) with an indication of their sizes. This mechanism is used when the size of the transmitted content is unknown in advance or can be very large, so the server sends the content in several fragments, each of which has a header indicating its size, and the last fragment has a size of 0. Also Transfer-Encoding: chunked can be used for real-time data transfer such as media streams where data is generated as it arrives. In this case, the client receives data fragments as soon as they appear, rather than waiting for the full download of all content.