HTTP protocol 6
Passing query string data in HTTP is used to pass parameters in the URL. The Query string is added to the URL after the ? and consists of one or more parameters that are passed in the key=value format. If there are several parameters, they are separated by the & symbol. When sending a request with parameters in the query string, they can be passed as part of the URL or in the request body, depending on the HTTP method. For the GET method, parameters are passed in the URL after the ? symbol. For methods POST, PUT and DELETE parameters are passed in the request body. In this case, you must specify a Content-Type header with a value of application/x-www-form-urlencoded to indicate that the data is being sent in URL-encoded format. When a request is received, the server may extract the parameters from the query string in the URL or from the request body, depending on the HTTP method, and process them appropriately.