HTTP protocol 7
HTTP redirect (HTTP Redirect) is a mechanism that allows you to redirect the user from one page of the site to another. The server can send an HTTP header with a 3XX status code indicating a redirect is needed and a new address to navigate to. The browser, after receiving such a header, automatically switches to a new address.
The most commonly used status codes for HTTP redirects are:
301 Moved Permanently - permanent redirect. This code indicates that the requested page has been permanently moved to a new address and the old address is no longer used. Browsers usually cache this type of redirect so that the next time the old address is accessed, the browser redirects to the new one on its own.
302 Found - temporary redirect. This code indicates that the requested page has been temporarily moved to a different address. At the same time, the old address is saved and can be used in the future.
307 Temporary Redirect - temporary redirect. Similar to 302 code, but more modern version.
308 Permanent Redirect - permanent redirect. Similar to code 301, but more modern version.
In addition, there are a few more status codes that can be used to redirect to HTTP, but they are less common.