HTTP protocol 8
Basic Authentication is a method of authenticating a user to a server by passing them a username and password as a base64 encoded string passed in the Authorization request header.

The authentication process with basic authentication is as follows:
The client sends a request to the server
The server returns a response with a 401 Unauthorized status code and a WWW-Authenticate header containing authentication parameters (for example, WWW-Authenticate: Basic realm="Restricted Area")
The client repeats the request with the Authorization header containing the login and password in base64 format (for example, Authorization: Basic QWxhZGRpbjpvcGVuIHNlc2FtZQ==)
When receiving a request, the server checks the correctness of the transmitted data, and if they are correct, responds to the request with the required response. If the username and password are incorrect, the server returns a 401 Unauthorized status code and requires re-authentication.