HTTP protocol 9
Cookies are small text files that are sent by the server to the client machine and stored there for use in subsequent requests. Cookies allow the server to "remember" information about the client, such as their preferences, session data, and other data.
Cookies are usually used for the following purposes:
Authentication and Authorization: The server can store user authentication information in cookies so that it does not require a username and password to be entered on every request.
Sessions: Cookies can be used to store information about the user's current session so that the server can save state between requests.
Personalization: Cookies allow the server to store information about user preferences and settings in order to provide personalized content.
Tracking User Behavior: Cookies can be used to track how users interact with the site, such as what pages they view and what products they buy.
Cookies are passed in HTTP request and response headers. The client can send cookies in requests using the Cookie header, and the server can set cookies in responses using the Set-Cookie header.
Cookies can be permanent or temporary. Permanent cookies are stored on the client machine longer than temporary cookies and are not deleted after the browser is closed. Temporary cookies are only stored until the browser is closed.
Cookies may also be secure or insecure. Secure cookies are only sent over an HTTPS connection and are not accessible to JavaScript, which helps prevent Cross-Site Scripting (XSS) and Cross-Site Request Forgery (CSRF) attacks. Insecure cookies can be used in XSS and CSRF attacks.