Sequence representation 2
Markup languages ​​are sets of rules for formatting text for the purpose of creating documentation, structuring information, and so on. In functional programming, the use of markup languages ​​is also very common. Markdown is a lightweight markup language that is widely used for creating documentation, blogs, and other text documents. It allows you to quickly and easily create headers, lists, tables, and more. Markdown can be translated to HTML, making it useful for creating web pages. HTML is the main markup language for creating web pages. HTML allows you to define the structure of a document, set headings, text blocks, images, links, and other elements. HTML can be used in conjunction with other languages ​​such as CSS and JavaScript to create beautiful, interactive and functional web pages. Functional programming uses libraries and tools to work with markup languages ​​such as Markdown and HTML that allow them to be treated as data rather than templates. This allows you to create documentation and web pages that can be easily manipulated and modified programmatically.