Create JS: Asynchronous programming 1
Asynchronous code is code that runs in the background and does not block the main thread of program execution. Instead, asynchronous code runs in parallel with other operations and can complete at any time.
JavaScript uses the concept of Promises and the async/await keywords to work with asynchronous code. Promises are objects that represent values ​​that may be available now, in the future, or never. The async keyword is used to create functions that run asynchronously. The await keyword is used to wait for an asynchronous operation to complete and receive the result.
Asynchronous code can be used to process requests to the server, download files, perform calculations in the background, and other operations that can take a long time. It can also improve application performance and provide a smoother user interface experience.