Create JS: Asynchronous programming 2
In asynchronous JavaScript code, there are several ways to return a value from an asynchronous function.
One way is to use callbacks. An asynchronous function takes a callback as an argument, which will be called when the operation completes. The result of the operation is passed to this callback as an argument.