Create Create JS: Asynchronous programming 3
In this lesson, we are talking about ordering asynchronous operations. According to the assignment, we need to implement and export the asynchronous compareFileSizes() function, which compares the sizes of two files and passes the result of the comparison to the passed callback function. If the first file is greater than the second, then it passes one, if the sizes are equal, then zero, otherwise -1.