Create JS: Asynchronous programming 6
In asynchronous programming, timers are used to perform certain operations at a given time or at a given interval. They allow you to schedule and run code in the future, as well as control delays and intervals between the execution of operations.

JavaScript has several functions for working with timers:
setTimeout: Starts the function after the specified time (in milliseconds). It takes two arguments: a function to execute and a delay time. Returns the ID of the timer.
setInterval: Runs the function after the specified time interval (in milliseconds). It takes two arguments: a function to execute and a time interval. Returns the ID of the timer.
clearTimeout and clearInterval: Stop the timer from running. They take as an argument the identifier of the timer obtained when the timer was created using setTimeout or setInterval.
Timers allow you to control asynchronous code execution, schedule operations, and respond to certain events or time delays. They are widely used in asynchronous programming to handle time delays, periodic operations, timeouts, and other scenarios.