Create JS: Asynchronous programming 7
Promises in JavaScript are an asynchronous programming mechanism that makes it easy to work with asynchronous operations and process their results.

A Promise is an object that represents the success or failure of an asynchronous operation. It has the following states:

Pending: The initial state of the promise. It is waiting for an asynchronous operation to complete.

Fulfilled: The promise enters this state when the asynchronous operation completes successfully. In this state, the promise contains the result of the operation.

Rejected: A promise enters this state when an asynchronous operation fails. In this state, the promise contains information about the error.

Promises allow you to chain method calls to sequentially execute asynchronous operations. This makes the code more readable and avoids the so-called "callback hell" (an anti-pattern where nested callbacks make the code harder to read and maintain).