Create JS: Asynchronous programming 8
In error handling in promises, you can use the catch() method to catch and handle errors. 
If the promise succeeds, then the callback passed to then() will be executed. In case of an error, the callback passed to catch() will be executed. Inside the catch() callback, you can perform the necessary actions to handle the error, such as displaying an error message or performing additional actions to recover from an error state.