Create JS: Asynchronous programming 13
The Axios library is a popular library for making HTTP requests in the browser and in the Node.js environment. It provides a simple and convenient interface for working with requests and responses, supports promises and provides the ability to intercept requests and responses.
You can make other types of requests such as POST, PUT, DELETE, etc. using the appropriate methods (axios.post, axios.put, axios.delete, etc.). You can specify request headers, send data in the request body, set request parameters, and so on.