JS: DOM API 2
DOM (Document Object Model) in JavaScript is a programming interface for interacting with HTML documents. It represents a document as a tree of objects, where each tree node represents a part of the document, such as elements, text nodes, or attributes.
With DOM in JavaScript, you can:
Access elements on the page and their attributes.
Create, modify, and delete elements and their content.
Manipulate styles and element classes.
React to events such as mouse clicks or key presses.
Change the structure of the document, add, move or remove nodes.