JS: DOM API 7
Events in JavaScript and the DOM API allow you to respond to user actions or web page changes. Events can be triggered by action actions, such as a mouse click, hover, button action, or a change in the value of a form element.
JavaScript and the DOM API have found many methods for dealing with events. Some of them:
-addEventListener: The addEventListener method allows you to include event handlers on a DOM element. You can specify the event type (for example, "click" or "key press") and a handler function to be called when the events occur.
-removeEventListener: The removeEventListener method is used to remove previously added event handlers. It takes the same parameters as addEventListener.
-event: When an event handler is called, an event object is passed to it. The event object contains information about the event, such as the event type, target element, pointer coordinates, and more. You can use the properties and methods of the event object to get more information or perform specific actions.