Create JS: DOM API 8
DOM API (Application Programming Interface) in JavaScript provides a set of methods and properties for working with DOM (Document Object Model) elements and handling events. Events allow you to respond to user actions or changes in the state of elements on the page. Here are some examples of working with events using the DOM API in JavaScript:

Adding an event handler:
Create a handleClick event handler function that prints a message to the console when the button is clicked. We then use the addEventListener method to add a click event handler to the button.
Removing an event handler:
If you need to remove a previously added event handler, you can use the removeEventListener method and pass in the same event type (click) and the same handler function (handleClick) that were used when adding the event handler.
Event object:
When an event handler is called, an event object is passed to it. This object contains information about the event and the element that fired the event. (event.target), that is, the element on which the event occurred, and the type of the event (event.type).
Prevent default action:
If you need to prevent the default action associated with an event (such as following a link), you can call the preventDefault() method on the event object.